+++ tests/base/test_boundBox.py
@@ -3,8 +3,10 @@ import sys
 import unittest
 
 thisFolder = os.path.dirname(os.path.abspath(__file__))
-parentFolder = os.path.dirname(thisFolder)
+baseFolder = os.path.dirname(thisFolder)
+parentFolder = os.path.dirname(baseFolder)
 sys.path.append(parentFolder)
+
 from liblathe.base.point import Point
 from liblathe.base.boundbox import BoundBox
