+++ tests/base/test_segment.py
@@ -4,8 +4,10 @@ import math
 import unittest
 
 thisFolder = os.path.dirname(os.path.abspath(__file__))
-parentFolder = os.path.dirname(thisFolder)
+baseFolder = os.path.dirname(thisFolder)
+parentFolder = os.path.dirname(baseFolder)
 sys.path.append(parentFolder)
+
 from liblathe.base.point import Point
 from liblathe.base.segment import Segment
 from liblathe.base.boundbox import BoundBox
