+++ tests/base/test_segmentGroup.py
@@ -3,7 +3,8 @@ import sys
 import unittest
 
 thisFolder = os.path.dirname(os.path.abspath(__file__))
-parentFolder = os.path.dirname(thisFolder)
+baseFolder = os.path.dirname(thisFolder)
+parentFolder = os.path.dirname(baseFolder)
 sys.path.append(parentFolder)
 
 from liblathe.base.point import Point
