+++ tests/op/test_rough_op.py
@@ -3,10 +3,11 @@ import sys
 import unittest
 
 thisFolder = os.path.dirname(os.path.abspath(__file__))
-parentFolder = os.path.dirname(thisFolder)
+opFolder = os.path.dirname(thisFolder)
+parentFolder = os.path.dirname(opFolder)
 sys.path.append(parentFolder)
 
-from liblathe.op.profile import ProfileOP
+from liblathe.op.rough import RoughOP
 from liblathe.base.command import Command
 from liblathe.base.boundbox import BoundBox
 from liblathe.base.point import Point
@@ -14,8 +15,8 @@ from liblathe.base.segment import Segment
 from liblathe.base.tool import Tool
 
 
-class test_ProfileOP(unittest.TestCase):
-    """Test for profile_op.py"""
+class test_RoughOP(unittest.TestCase):
+    """Test for rough_op.py"""
 
     def setUp(self):
 
@@ -26,26 +27,22 @@ class test_ProfileOP(unittest.TestCase):
         PartPt2 = Point(-5, 0, -9)
         PartPt3 = Point(-9.5, 0, -15.85)
         PartPt4 = Point(-5.4, 0, -22)
-        PartPt5 = Point(-5.4, 0, -30)
-        PartPt6 = Point(-5.4, 0, -35)
-        PartPt7 = Point(-5.4, 0, -40)
-        PartPt8 = Point(-13, 0, -45)
-        PartPt9 = Point(-13, 0, -48)
-        PartPt10 = Point(0, 0, -48)
+        PartPt5 = Point(-5.4, 0, -40)
+        PartPt6 = Point(-13, 0, -45)
+        PartPt7 = Point(-13, 0, -48)
+        PartPt8 = Point(0, 0, -48)
 
         part_segments.append(Segment(PartPt1, PartPt2, -0.75))
         part_segments.append(Segment(PartPt2, PartPt3))
         part_segments.append(Segment(PartPt3, PartPt4))
-        part_segments.append(Segment(PartPt4, PartPt5, 0.25))
+        part_segments.append(Segment(PartPt4, PartPt5))
         part_segments.append(Segment(PartPt5, PartPt6))
         part_segments.append(Segment(PartPt6, PartPt7))
         part_segments.append(Segment(PartPt7, PartPt8))
-        part_segments.append(Segment(PartPt8, PartPt9))
-        part_segments.append(Segment(PartPt9, PartPt10))
 
         # Define stock bounds
         stockPt1 = Point(0, 0, 15)
-        stockPt2 = Point(-25, 0, -55)
+        stockPt2 = Point(-25, 0, -47.5)
         stock_boundbox = BoundBox(stockPt1, stockPt2)
 
         # set feed rate to test
@@ -59,12 +56,12 @@ class test_ProfileOP(unittest.TestCase):
         params['end_offset'] = 0
         params['allow_grooving'] = True
         params['step_over'] = 1
-        params['finish_passes'] = 3
+        params['finish_passes'] = 2
         params['stock_to_leave'] = 0
         params['hfeed'] = self.hfeed
         params['vfeed'] = 10
 
-        self.op = ProfileOP()
+        self.op = RoughOP()
         self.op.set_params(params)
         self.op.add_stock(stock_boundbox)
         self.op.add_part_edges(part_segments)
@@ -80,7 +77,12 @@ class test_ProfileOP(unittest.TestCase):
         # check there is a return list
         self.assertTrue(len(gcode))
 
+        # get the minimum z value
+        min_z = 0
+
         for command in gcode:
+
+            print(command.get_movement(), command.get_params())
 
             # check the command is a liblathe.command.Command
             self.assertTrue(isinstance(command, Command))
@@ -96,6 +98,13 @@ class test_ProfileOP(unittest.TestCase):
                 # check the feed rate
                 if feed_exists:
                     self.assertEqual(command.get_params()['F'], self.hfeed)
+                
+                # capture the min z value
+                if "Z" in command.get_params():
+                    min_z = min(command.get_params()['Z'], min_z)
+
+        # test the z value matches the stock z_min
+        self.assertEqual(min_z, self.op.stock.z_min)
 
 
 if __name__ == '__main__':
